refactor: выносит блокировку основной формы из js в html для более раннего срабатывания
