+++ js/upload/main.js
@@ -1,12 +1,19 @@
-import { activateForm, deactivateForm } from './activation.js';
 import { checkValidity, resetValidity } from './validation.js';
 import { activateSlider, deactivateSlider, resetSlider } from './price.js';
 
-deactivateForm();
-deactivateSlider();
-
 const form = document.querySelector('.ad-form');
 const submitButton = document.querySelector('.ad-form__submit');
+const fieldsets = form.querySelectorAll('fieldset');
+
+deactivateSlider();
+
+const activateForm = () => {
+  form.classList.remove('ad-form--disabled');
+  fieldsets.forEach((fieldset) => {
+    fieldset.disabled = false;
+  });
+  activateSlider();
+};
 
 const setSubmitDisabled = (flag) => {
   submitButton.disabled = flag;
@@ -29,4 +36,4 @@ const resetUpload = () => {
   resetSlider();
 };
 
-export { setSubmitDisabled, activateForm, activateSlider, setCoordinates, resetUpload };
+export { setSubmitDisabled, activateForm, setCoordinates, resetUpload };
